Add support for sorting 'FRS/GMRS' channel names.
Support has been added to sort channels with names that start with 'FRS/GMRS',
followed by a space, followed by a channel number, to the end of the file,
ordered by the channel number.

Without this, the FRS and GMRS channels were ordered by frequency, which
resulted in an odd channel order.

This could easily be extended to other radio services that use channels in
the future.